pyproject.toml
[project]
name = "pp"
version = "0.1.0"
description = "PowerPoint Inverter - Invert colors in PowerPoint presentations"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "python-pptx>=1.0.2",
    "streamlit>=1.52.1",
    "pillow>=10.0.0",
    "numpy>=2.2.6",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
pp = "pp.app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pp"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
